"Entities changes"

+++ efka-jpa/src/main/java/org/regeneration/efkajpa/repository/SpecialtiesRepository.java
@@ -3,5 +3,5 @@ package org.regeneration.efkajpa.repository;
 import org.regeneration.efkajpa.entity.Specialties;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpecialtiesRepository extends JpaRepository<Specialties, Long> {
+public interface SpecialtiesRepository extends JpaRepository<Specialties, Integer> {
 }
